+++ src/components/date-range/index.ts
@@ -1,3 +1,3 @@
 export { default } from "./date-range.component";
 export type { DateRangeChangeEvent } from "./date-range.component";
-export type { DateRangeContextProps } from "./date-range.context";
+export type { DateRangeContextProps } from "./__internal__/date-range.context";
